requirements.txt
requests>=2.20.0
retrying==1.3.3
rpi-ws281x==4.1.0
xmltodict==0.11.0
RPi.GPIO==0.6.5
python-crontab==2.3.5
